requirements-mac.txt
tensorflow-macos==2.15.0
tensorflow-metal==1.1.0
numpy==1.26.4
pandas==2.2.2
matplotlib==3.8.4
scikit-learn==1.5.1
lightkurve==2.4.2
joblib==1.4.2
streamlit==1.39.0
astropy==6.1.3
scipy==1.13.1
